moved server.js into api folder, consolidated app.js and server.js, need to do more debugging
