+++ api/package.json
@@ -1,17 +1,29 @@
 {
-  "name": "api",
-  "version": "0.0.0",
+  "name": "sentinel",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
+    "start:prod": "node server.js",
+    "start:dev": "concurrently \"node server.js\" \"cd ../ && expo start --android\"",
+    "client": "cd ../ && expo start --android",
+    "seed": "node scripts/seedsDB.js",
+    "install": "cd ../ && npm install",
+    "build": "cd ../ && npm run build",
+    "heroku-postbuild": "cd ../ && npm install && npm install --only=dev --no-shrinkwrap && npm run build"
+  },
+  "devDependencies": {
+    "concurrently": "^5.1.0"
   },
   "dependencies": {
+    "axios": "^0.19.2",
+    "bcrypt": "^4.0.0",
     "cookie-parser": "~1.4.4",
     "cors": "^2.8.5",
     "debug": "~2.6.9",
     "express": "~4.16.1",
     "http-errors": "~1.6.3",
-    "jade": "~1.11.0",
+    "if-env": "^1.0.4",
     "jsonwebtoken": "^8.5.1",
     "morgan": "~1.9.1"
   }
